+++ frontend/src/metabase/query_builder/components/ExpressionEditor/HelpText/index.js
@@ -1,0 +1,1 @@
+export { default } from "./HelpText";
